backend/requirements.txt
# AirAware Backend - Django Requirements
# Core Django
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1

# Database
psycopg2-binary>=2.9.7
django-extensions>=3.2.3

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
cryptography>=41.0.7

# API Documentation
drf-yasg>=1.21.7
drf-spectacular>=0.26.5

# Environment & Configuration
python-decouple>=3.8
django-environ>=0.11.2

# Air Quality Data APIs
requests>=2.31.0
httpx>=0.25.2
aiohttp>=3.9.1

# Data Processing & Analysis
pandas>=2.1.4
numpy>=1.25.2
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0

# Geospatial & Mapping
geopy>=2.4.0
folium>=0.15.0
geopandas>=0.14.1
shapely>=2.0.2

# Time Series & Weather Data
pytz>=2023.3
python-dateutil>=2.8.2
timezonefinder>=6.2.0

# NASA & Satellite Data Integration
nasapy>=0.0.2
h5py>=3.10.0
netCDF4>=1.6.5

# CPCB & Government API Integration
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2

# Caching & Performance
redis>=5.0.1
django-redis>=5.4.0
celery>=5.3.4

# Monitoring & Logging
sentry-sdk>=1.38.0
django-health-check>=3.17.0

# Email Services
django-anymail>=10.2

# Image Processing (for satellite imagery)
Pillow>=10.1.0
opencv-python>=4.8.1.78

# Development & Testing
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0
coverage>=7.3.2
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
django-storages>=1.14.2
boto3>=1.34.0

# Real-time Features
channels>=4.0.0
channels-redis>=4.1.0
websockets>=12.0